Moved code around and cleaned up includes.
