+++ Embedded/MaxMix/Communications.h
@@ -1,7 +1,6 @@
 #pragma once
 
-#include <Config.h>
-#include <Communications/Messages.h>
+#include "Config.h"
 
 namespace Communications
 {
